move tab component to separate module
