+++ src/renderer/screens/Tabs/types.ts
@@ -1,0 +1,7 @@
+// Copyright 2024 @polkadot-live/polkadot-live-app authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+export interface TabProps {
+  id: number;
+  label: string;
+}
